refactor: sanitizeName in ein gemeinsames Modul
